Tests/LibWeb/Text/expected/css/style-invalidation/viewport-resize-layered-import-broad-invalidation.txt

initial layered import color: rgb(255, 0, 0)
resized layered import color: rgb(0, 128, 0)
resized back layered import color: rgb(255, 0, 0)
